pyproject.toml
[project]
name = "minirouter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "evdev>=1.9.2",
    "pillow>=12.1.0",
    "python-statemachine>=2.5.0",
    "pyzmq>=27.1.0",
    "requests>=2.32.5",
    "sdbus-networkmanager>=2.0.0",
]

[project.scripts]
minirouter = "minirouter.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["minirouter"]

[tool.ruff]
line-length = 120
lint.unfixable = [
  "F401", # Don't touch unused imports
]
